+++ tools/eMMC/bbb-eMMC-flasher-eewiki-ext4.sh
@@ -24,8 +24,8 @@
 #This script assumes, these packages are installed, as network may not be setup
 #dosfstools initramfs-tools rsync u-boot-tools
 
-version_message="1.20160909: u-boot 1MB -> 4MB hole..."
-
+version_message="1.20161005: sfdisk: actually calculate the start of 2nd/3rd partitions..."
+#
 #https://rcn-ee.com/repos/bootloader/am335x_evm/
 http_spl="MLO-am335x_evm-v2016.03-r7"
 http_uboot="u-boot-am335x_evm-v2016.03-r7.img"
@@ -63,16 +63,29 @@ inf_loop () {
 	done
 }
 
+get_device () {
+	is_bbb="enable"
+	machine=$(cat /proc/device-tree/model | sed "s/ /_/g")
+
+	case "${machine}" in
+	TI_AM5728_BeagleBoard*)
+		unset is_bbb
+		;;
+	esac
+}
+
 write_failure () {
 	message="writing to [${destination}] failed..." ; broadcast
 
-	[ -e /proc/$CYLON_PID ]  && kill $CYLON_PID > /dev/null 2>&1
+	if [ "x${is_bbb}" = "xenable" ] ; then
+		[ -e /proc/$CYLON_PID ]  && kill $CYLON_PID > /dev/null 2>&1
 
-	if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
-		echo heartbeat > /sys/class/leds/beaglebone\:green\:usr0/trigger
-		echo heartbeat > /sys/class/leds/beaglebone\:green\:usr1/trigger
-		echo heartbeat > /sys/class/leds/beaglebone\:green\:usr2/trigger
-		echo heartbeat > /sys/class/leds/beaglebone\:green\:usr3/trigger
+		if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
+			echo heartbeat > /sys/class/leds/beaglebone\:green\:usr0/trigger
+			echo heartbeat > /sys/class/leds/beaglebone\:green\:usr1/trigger
+			echo heartbeat > /sys/class/leds/beaglebone\:green\:usr2/trigger
+			echo heartbeat > /sys/class/leds/beaglebone\:green\:usr3/trigger
+		fi
 	fi
 	message="-----------------------------" ; broadcast
 	flush_cache
@@ -113,6 +126,7 @@ check_eeprom () {
 		got_eeprom="true"
 	fi
 
+	if [ "x${is_bbb}" = "xenable" ] ; then
 		if [ "x${got_eeprom}" = "xtrue" ] ; then
 			eeprom_header=$(hexdump -e '8/1 "%c"' ${eeprom} -n 8 | cut -b 6-8)
 			if [ "x${eeprom_header}" = "x335" ] ; then
@@ -141,6 +155,7 @@ check_eeprom () {
 				fi
 			fi
 		fi
+	fi
 }
 
 check_running_system () {
@@ -168,62 +183,58 @@ check_running_system () {
 	fi
 	flush_cache
 
-	##FIXME: quick check for rsync 3.1 (jessie)
-	unset rsync_check
-	unset rsync_progress
-	rsync_check=$(LC_ALL=C rsync --version | grep version | awk '{print $3}' || true)
-	if [ "x${rsync_check}" = "x3.1.1" ] ; then
-		rsync_progress="--info=progress2 --human-readable"
-	fi
-
-	if [ ! -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
-		modprobe leds_gpio || true
-		sleep 1
+	if [ "x${is_bbb}" = "xenable" ] ; then
+		if [ ! -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
+			modprobe leds_gpio || true
+			sleep 1
+		fi
 	fi
 }
 
 cylon_leds () {
-	if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
-		BASE=/sys/class/leds/beaglebone\:green\:usr
-		echo none > ${BASE}0/trigger
-		echo none > ${BASE}1/trigger
-		echo none > ${BASE}2/trigger
-		echo none > ${BASE}3/trigger
+	if [ "x${is_bbb}" = "xenable" ] ; then
+		if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
+			BASE=/sys/class/leds/beaglebone\:green\:usr
+			echo none > ${BASE}0/trigger
+			echo none > ${BASE}1/trigger
+			echo none > ${BASE}2/trigger
+			echo none > ${BASE}3/trigger
 
-		STATE=1
-		while : ; do
-			case $STATE in
-			1)	echo 255 > ${BASE}0/brightness
-				echo 0   > ${BASE}1/brightness
-				STATE=2
-				;;
-			2)	echo 255 > ${BASE}1/brightness
-				echo 0   > ${BASE}0/brightness
-				STATE=3
-				;;
-			3)	echo 255 > ${BASE}2/brightness
-				echo 0   > ${BASE}1/brightness
-				STATE=4
-				;;
-			4)	echo 255 > ${BASE}3/brightness
-				echo 0   > ${BASE}2/brightness
-				STATE=5
-				;;
-			5)	echo 255 > ${BASE}2/brightness
-				echo 0   > ${BASE}3/brightness
-				STATE=6
-				;;
-			6)	echo 255 > ${BASE}1/brightness
-				echo 0   > ${BASE}2/brightness
-				STATE=1
-				;;
-			*)	echo 255 > ${BASE}0/brightness
-				echo 0   > ${BASE}1/brightness
-				STATE=2
-				;;
-			esac
-			sleep 0.1
-		done
+			STATE=1
+			while : ; do
+				case $STATE in
+				1)	echo 255 > ${BASE}0/brightness
+					echo 0   > ${BASE}1/brightness
+					STATE=2
+					;;
+				2)	echo 255 > ${BASE}1/brightness
+					echo 0   > ${BASE}0/brightness
+					STATE=3
+					;;
+				3)	echo 255 > ${BASE}2/brightness
+					echo 0   > ${BASE}1/brightness
+					STATE=4
+					;;
+				4)	echo 255 > ${BASE}3/brightness
+					echo 0   > ${BASE}2/brightness
+					STATE=5
+					;;
+				5)	echo 255 > ${BASE}2/brightness
+					echo 0   > ${BASE}3/brightness
+					STATE=6
+					;;
+				6)	echo 255 > ${BASE}1/brightness
+					echo 0   > ${BASE}2/brightness
+					STATE=1
+					;;
+				*)	echo 255 > ${BASE}0/brightness
+					echo 0   > ${BASE}1/brightness
+					STATE=2
+					;;
+				esac
+				sleep 0.1
+			done
+		fi
 	fi
 }
 
@@ -320,7 +331,7 @@ copy_boot () {
 	flush_cache
 	umount /tmp/boot/ || umount -l /tmp/boot/ || write_failure
 	flush_cache
-	umount /boot/uboot || umount -l /boot/uboot
+	umount /boot/uboot || umount -l /boot/uboot || true
 }
 
 copy_rootfs () {
@@ -380,7 +391,9 @@ copy_rootfs () {
 	flush_cache
 	umount /tmp/rootfs/ || umount -l /tmp/rootfs/ || write_failure
 
-	[ -e /proc/$CYLON_PID ]  && kill $CYLON_PID
+	if [ "x${is_bbb}" = "xenable" ] ; then
+		[ -e /proc/$CYLON_PID ]  && kill $CYLON_PID
+	fi
 
 	message="Syncing: ${destination}" ; broadcast
 	#https://github.com/beagleboard/meta-beagleboard/blob/master/contrib/bone-flash-tool/emmc.sh#L158-L159
@@ -396,16 +409,19 @@ copy_rootfs () {
 		message="debug: enabled" ; broadcast
 		inf_loop
 	else
-		if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
-			echo default-on > /sys/class/leds/beaglebone\:green\:usr0/trigger
-			echo default-on > /sys/class/leds/beaglebone\:green\:usr1/trigger
-			echo default-on > /sys/class/leds/beaglebone\:green\:usr2/trigger
-			echo default-on > /sys/class/leds/beaglebone\:green\:usr3/trigger
+		if [ "x${is_bbb}" = "xenable" ] ; then
+			if [ -e /sys/class/leds/beaglebone\:green\:usr0/trigger ] ; then
+				echo default-on > /sys/class/leds/beaglebone\:green\:usr0/trigger
+				echo default-on > /sys/class/leds/beaglebone\:green\:usr1/trigger
+				echo default-on > /sys/class/leds/beaglebone\:green\:usr2/trigger
+				echo default-on > /sys/class/leds/beaglebone\:green\:usr3/trigger
+			fi
 		fi
 		mount
 
 		message="eMMC has been flashed: please wait for device to power down." ; broadcast
 		message="-----------------------------" ; broadcast
+
 		flush_cache
 	fi
 }
@@ -497,23 +513,25 @@ partition_drive () {
 
 		sfdisk_options="--force --Linux --in-order --unit M"
 		sfdisk_boot_startmb="${conf_boot_startmb}"
-		sfdisk_boot_endmb="${conf_boot_endmb}"
+		sfdisk_boot_size_mb="${conf_boot_endmb}"
+		sfdisk_rootfs_startmb=$(($sfdisk_boot_startmb + $sfdisk_boot_size_mb))
 
 		test_sfdisk=$(LC_ALL=C sfdisk --help | grep -m 1 -e "--in-order" || true)
 		if [ "x${test_sfdisk}" = "x" ] ; then
 			message="sfdisk: [2.26.x or greater]" ; broadcast
 			sfdisk_options="--force"
 			sfdisk_boot_startmb="${sfdisk_boot_startmb}M"
-			sfdisk_boot_endmb="${sfdisk_boot_endmb}M"
+			sfdisk_boot_size_mb="${sfdisk_boot_size_mb}M"
+			sfdisk_rootfs_startmb="${sfdisk_rootfs_startmb}M"
 		fi
 
 		message="sfdisk: [sfdisk ${sfdisk_options} ${destination}]" ; broadcast
-		message="sfdisk: [${sfdisk_boot_startmb},${sfdisk_boot_endmb},${sfdisk_fstype},*]" ; broadcast
-		message="sfdisk: [,,,-]" ; broadcast
+		message="sfdisk: [${sfdisk_boot_startmb},${sfdisk_boot_size_mb},${sfdisk_fstype},*]" ; broadcast
+		message="sfdisk: [${sfdisk_rootfs_startmb},,,-]" ; broadcast
 
 		LC_ALL=C sfdisk ${sfdisk_options} "${destination}" <<-__EOF__
-			${sfdisk_boot_startmb},${sfdisk_boot_endmb},${sfdisk_fstype},*
-			,,,-
+			${sfdisk_boot_startmb},${sfdisk_boot_size_mb},${sfdisk_fstype},*
+			${sfdisk_rootfs_startmb},,,-
 		__EOF__
 
 		flush_cache
@@ -551,7 +569,7 @@ partition_drive () {
 
 		message="sfdisk: [$(LC_ALL=C sfdisk --version)]" ; broadcast
 		message="sfdisk: [sfdisk ${sfdisk_options} ${destination}]" ; broadcast
-		message="sfdisk: [${sfdisk_boot_startmb},${sfdisk_boot_endmb},${sfdisk_fstype},*]" ; broadcast
+		message="sfdisk: [${sfdisk_boot_startmb},,${sfdisk_fstype},*]" ; broadcast
 
 		LC_ALL=C sfdisk ${sfdisk_options} "${destination}" <<-__EOF__
 			${sfdisk_boot_startmb},,${sfdisk_fstype},*
@@ -572,6 +590,7 @@ message="-----------------------------" ; broadcast
 message="Version: [${version_message}]" ; broadcast
 message="-----------------------------" ; broadcast
 
+get_device
 check_eeprom
 check_running_system
 cylon_leds & CYLON_PID=$!
